genai/requirements.txt
openai
python-dotenv
fastapi
uvicorn
mysql-connector-python